Added alternative prime generation method to Prime Factorization project
I wanted to implement the Sieve of Erastothenes
algorithm to generate primes. It actually performs worse
basically because I can't use it as a generator so it
creates a whole bunch of primes even though not all of
them might be used in the factorization.

Still, worked as an exercise in writing Python :)

I left the code using the initial naive prime
generation method.